perf(jsonSchema & withCompuListHighOrder & etc): 调整代码结构
调整withCompuListHighOrder代码结构，将同类型属性列表的计算移动到jsonSchema文件,调整各类型的creator
