+++ src/components/SchemaCreator/ArraySchemaCreator.js
@@ -86,23 +86,23 @@ class ArraySchemaCreator extends React.Component {
 
   componentWillReceiveProps (nextProps) {
     console.log('o nextProps', nextProps);
-    let res = nextProps.compuListPrepare(nextProps);
+    // let res = nextProps.compuListPrepare(nextProps);
     this.setState({
       ownerList: nextProps.ownerList ? nextProps.ownerList : [],
       defList: nextProps.defList ? nextProps.defList : [],
       refList: nextProps.refList ? nextProps.refList : [],
-      arrayTypeList: res.sameTypeListObj.tmpArrayList
+      arrayTypeList: nextProps.arrayTypeList ? nextProps.arrayTypeList : []
     });
   }
 
   componentDidMount () {
     console.log('o properties: ', this.props.properties);
-    let res = this.props.compuListPrepare(this.props);
+    // let res = this.props.compuListPrepare(this.props);
     this.setState({
       ownerList: this.props.ownerList ? this.props.ownerList : [],
       defList: this.props.defList ? this.props.defList : [],
       refList: this.props.refList ? this.props.refList : [],
-      arrayTypeList: res.sameTypeListObj.tmpArrayList
+      arrayTypeList: this.props.arrayTypeList ? this.props.arrayTypeList : []
     });
   }
